feat(ui): new parameters view
