+++ web-src/src/App.css
@@ -56,3 +56,7 @@ main article {
 .error {
   background-color: red;
 }
+
+.success {
+  background-color: green;
+}
