+++ web-src/src/App.js
@@ -9,38 +9,28 @@ import ErrorBoundary from 'react-error-boundary'
 import moment from 'moment'
 
 import {
-  ActionButton,
   Button,
   Content,
   defaultTheme as theme,
-  Dialog,
-  DialogTrigger,
-  Divider,
-  Flex,
   Heading,
   IllustratedMessage, 
   Link,
   ProgressCircle,
   Provider,
+  Radio,
+  RadioGroup,
   TextField,
   View,
   Well 
 } from '@adobe/react-spectrum'
-import { 
-  Cell,
-  Column,
-  Row,
-  Table,
-  TableBody,
-  TableHeader
-} from '@react-spectrum/table'
 
 import './App.css'
 
 import NotFound from '@spectrum-icons/illustrations/NotFound';
 import Error from '@spectrum-icons/illustrations/Error';
-import espagonLogo from '../resources/epsagon.svg'
-import coralogixLogo from '../resources/coralogix.png'
+
+import ViewRequests from './ViewRequests'
+import ViewParameters from './ViewParameters'
 
 import { actionWebInvoke } from './utils'
 
@@ -62,6 +52,7 @@ export default class App extends React.Component {
     )
 
     this.handleIdChange = this.handleIdChange.bind(this)
+    this.handleViewChange = this.handleViewChange.bind(this)
     this.handleKeyDown = this.handleKeyDown.bind(this)
     this.componentDidMount = this.componentDidMount.bind(this)
     
@@ -76,7 +67,9 @@ export default class App extends React.Component {
     }
 
     this.state = {
-      id
+      id,
+      view: 'requests'
+      // view: 'params'
     }
 
     console.debug('runtime object:', this.props.runtime)
@@ -131,63 +124,15 @@ export default class App extends React.Component {
     window.location.href = u.href
   }
 
+  handleViewChange(value) {
+    this.setState({'view': value})
+  }
+
   handleKeyDown(evt) {
     if (evt.keyCode === 13) {
       // enter key pressed
       this.search()
     }
-  }
-
-  getViewParams(params) {
-    if (params) {
-      const list = Object.keys(params).map((k) => {
-        const v = params[k]
-        if (!(v instanceof Object)) {
-          return (
-            <li key={k}>{`${k}: ${v}`}</li>
-          )
-        } else {
-          return (
-            <li key={k}>{`${k}:`}{this.getViewParams(v)}</li>
-          )
-        }
-        
-      })
-      return <ul>{list}</ul>
-    }
-    return <span>No data</span>
-  }
-
-  getViewResponse(response) {
-    if (response) {
-      const list = Object.keys(response).map((k) => {
-        const v = response[k]
-        if (!(v instanceof Object)) {
-          return (
-            <li key={k}>{`${k}: ${v}`}</li>
-          )
-        } else {
-          return (
-            <li key={k}>{`${k}:`}{this.getViewResponse(v)}</li>
-          )
-        }
-        
-      })
-      return <ul>{list}</ul>
-    }
-    return <span>No data</span>
-  }
-
-  getViewLogs(logs) {
-    if (logs) {
-      const list = logs.map((entry, index) => {
-        return (
-          <li key={index}>{`${entry.date}: ${entry.message}`}</li>
-        )
-      })
-      return <ul>{list}</ul>
-    }
-    return <span>No logs</span>
   }
 
   search() {
@@ -209,19 +154,7 @@ export default class App extends React.Component {
     return moment(date).format('LL')
   }
 
-  viewUTCTime(date) {
-    return moment(date).utc().format('HH:mm:ss.SSS')
-  }
-
-  createIndent(level) {
-    const indent = [];
-    for(let i=0;i<level;i++) {
-      indent.push(<span key={i} className="indentBox"></span>)
-    }
-    return indent;
-  }
-
-  render () {
+   render () {
     return (
       // ErrorBoundary wraps child components to handle eventual rendering errors
       <ErrorBoundary onError={ this.onError } FallbackComponent={ this.fallbackComponent } >
@@ -288,144 +221,19 @@ export default class App extends React.Component {
                       <span><b>URL:</b> <Link variant="primary"><a href={this.state.spans[0].url}>{this.state.spans[0].url}</a></Link></span>
                     </p>
                   </Content>
-                  <Table aria-label="Results">
-                    <TableHeader>
-                      <Column width="7%">Time (UTC)</Column>
-                      <Column width="25%">Action</Column>
-                      <Column width="16%">Activation ID</Column>
-                      <Column width="12%">Path</Column>
-                      <Column width="4%">Status</Column>
-                      <Column width="4%"><img className="custom-icon" src={espagonLogo} alt="View in Epsagon" title="View in Epsagon"/></Column>
-                      <Column width="4%"><img className="custom-icon" src={coralogixLogo} alt="View in Coralogix" title="View in Coralogix"/></Column>
-                      <Column width="7%">Response</Column>
-                      <Column width="7%">Logs</Column>
-                      <Column width="7%">Replay</Column>
-                      <Column width="7%">Data</Column>
-                    </TableHeader>
-                    <TableBody>
-                      { this.state.spans.map((span, index) => {
-                        if (span.invisible) return;
-                        const espagonLink = `https://dashboard.epsagon.com/spans/${span.spanId}`
-                        const coralogixLink = `https://helix.coralogix.com/#/query/logs?query=${span.activationId || this.state.id}`
 
-                        const hasData = span.data && Object.keys(span.data).length > 0
-                        const dataButtonLabel = `Data`
-
-                        const hasResponse = !!span.response
-
-                        const hasLogs = span.logs && span.logs.length > 0
-                        const logsButtonLabel = `${hasLogs ? span.logs.length : ''} logs`
-
-                        const errorClassName = span.error || span.status >= 500 ? 'error' : ''
-
-                        let replayURL = ''
-                        let canReplay = false
-                        if (span.host) {
-                          if (span.invokedName) {
-                            canReplay = true
-                            const u = new URL(`${span.host}/api/v1/web${span.invokedName}`)
-                            if (span.params) {
-                              for(let p in span.params) {
-                                u.searchParams.append(p, span.params[p]);
-                              }
-                            }
-                            replayURL = u.toString();
-                          } else {
-                            if (span.path) {
-                              canReplay = true
-                              const u = new URL(`${span.host.indexOf('https') === 0 ? span.host : 'https://' + span.host}${span.path}`)
-                              if (span.params) {
-                                for(let p in span.params) {
-                                  u.searchParams.append(p, span.params[p]);
-                                }
-                              }
-                              replayURL = u.toString();
-                            }
-                          }
-                        } else {
-                          if (span.name === 'fastly') {
-                            canReplay = true
-                            replayURL = span.url
-                          }
-                        }
-                        return <Row key={index} UNSAFE_className={errorClassName}>
-                          <Cell>{this.viewUTCTime(span.date)}</Cell>
-                          <Cell>
-                            <span className="indentOffset">{this.createIndent(span.level)}</span>
-                            <span className="spanName" title={span.invokedName || span.name}>{span.invokedName || span.name}</span>
-                          </Cell>
-                          <Cell>{span.activationId}</Cell>
-                          <Cell><span className="pathCell" title={span.path}>{span.path}</span></Cell>
-                          <Cell>{span.status}</Cell>
-                          <Cell>  { span.spanId &&
-                            <a href={espagonLink} target="_new"><img className="custom-icon" src={espagonLogo} alt="View in Epsagon" title="View in Epsagon"/></a>
-                          }
-                          </Cell>
-                          <Cell><a href={coralogixLink} target="_new"><img className="custom-icon" src={coralogixLogo} alt="View in Coralogix" title="View in Coralogix"/></a></Cell>
-                          <Cell> { hasResponse &&
-                              <DialogTrigger type="fullscreenTakeover">
-                                <ActionButton>Response</ActionButton>
-                                {(close) => (
-                                  <Dialog>
-                                    <Heading>Response</Heading>
-                                    <Divider />
-                                    <Content>
-                                      <Flex direction="column">
-                                        <div>{ this.getViewResponse(span.response) }</div>
-                                        <Button alignSelf="center" variant="cta" onPress={close}>Close</Button>
-                                      </Flex>
-                                    </Content>
-                                  </Dialog>
-                                )}
-                              </DialogTrigger>
-                          }
-                          </Cell>
-                          <Cell> 
-                            { hasLogs && 
-                            <DialogTrigger type="fullscreenTakeover">
-                              <ActionButton>{logsButtonLabel}</ActionButton>
-                              {(close) => (
-                                <Dialog>
-                                  <Heading>Action logs</Heading>
-                                  <Divider />
-                                  <Content>
-                                    <Flex direction="column">
-                                      <div>{ this.getViewLogs(span.logs) }</div>
-                                      <Button alignSelf="center" variant="cta" onPress={close}>Close</Button>
-                                    </Flex>
-                                  </Content>
-                                </Dialog>
-                              )}
-                            </DialogTrigger>
-                          }
-                          </Cell>
-                          <Cell> { canReplay && 
-                            <ActionButton onClick={() => { window.open(replayURL)} }>Replay</ActionButton>
-                          }
-                          </Cell>
-                          <Cell> 
-                            { hasData && 
-                            <DialogTrigger type="fullscreenTakeover">
-                              <ActionButton>{dataButtonLabel}</ActionButton>
-                              {(close) => (
-                                <Dialog>
-                                  <Heading>Data</Heading>
-                                  <Divider />
-                                  <Content>
-                                    <Flex direction="column">
-                                      <div>{ this.getViewParams(span.data) }</div>
-                                      <Button alignSelf="center" variant="cta" onPress={close}>Close</Button>
-                                    </Flex>
-                                  </Content>
-                                </Dialog>
-                              )}
-                            </DialogTrigger>
-                          }
-                          </Cell>
-                        </Row>
-                      })}
-                    </TableBody>
-                  </Table>
+                  <RadioGroup label="View" orientation="horizontal" value={this.state.view} onChange={this.handleViewChange} >
+                    <Radio value="requests">Requests</Radio>
+                    <Radio value="params">Parameters</Radio>
+                  </RadioGroup>
+                  {
+                    this.state.view === 'requests' &&
+                      <ViewRequests id={this.state.id} spans={this.state.spans}/>
+                  }
+                  {
+                    this.state.view === 'params' &&
+                      <ViewParameters id={this.state.id} spans={this.state.spans}/>
+                  }
                   <br/>
                   <br/>
                 </div>
